backend/chat_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
openai==1.3.7
tiktoken==0.5.2
